clean up stats for shapefile
